+++ source/rafcon/gui/controllers/utils/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2016-2017 DLR
+# Copyright (C) 2014-2017 DLR
 #
 # All rights reserved. This program and the accompanying materials are made
 # available under the terms of the Eclipse Public License v1.0 which
@@ -6,6 +6,8 @@
 # http://www.eclipse.org/legal/epl-v10.html
 #
 # Contributors:
+# Annika Wollschlaeger <annika.wollschlaeger@dlr.de>
+# Franz Steinmetz <franz.steinmetz@dlr.de>
 # Rico Belder <rico.belder@dlr.de>
 # Sebastian Brunner <sebastian.brunner@dlr.de>