+++ source/rafcon/gui/models/container_state.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2014-2017 DLR
+# Copyright (C) 2014-2018 DLR
 #
 # All rights reserved. This program and the accompanying materials are made
 # available under the terms of the Eclipse Public License v1.0 which
@@ -140,6 +140,12 @@ class ContainerStateModel(StateModel):
         for state_element in sorted(self.states.values()) + sorted(self.transitions[:] + self.data_flows[:] + \
                                                                    self.scoped_variables[:]):
             self.update_hash_from_dict(obj_hash, state_element)
+
+    def update_meta_data_hash(self, obj_hash):
+        super(ContainerStateModel, self).update_meta_data_hash(obj_hash)
+        for state_element in sorted(self.states.values()) + sorted(self.transitions[:] + self.data_flows[:] + \
+                                                                   self.scoped_variables[:]):
+            state_element.update_meta_data_hash(obj_hash)
 
     @ModelMT.observe("state", before=True, after=True)
     def model_changed(self, model, prop_name, info):