+++ source/rafcon/gui/models/state_machine.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2015-2017 DLR
+# Copyright (C) 2015-2018 DLR
 #
 # All rights reserved. This program and the accompanying materials are made
 # available under the terms of the Eclipse Public License v1.0 which
@@ -25,6 +25,7 @@ from rafcon.core.states.container_state import ContainerState
 from rafcon.core.states.library_state import LibraryState
 from rafcon.core.storage import storage
 from rafcon.gui.config import global_gui_config
+from rafcon.gui.models.meta import MetaModel
 from rafcon.gui.models import ContainerStateModel, StateModel, LibraryStateModel
 from rafcon.gui.models.selection import Selection
 from rafcon.gui.models.signals import MetaSignalMsg
@@ -36,7 +37,7 @@ from rafcon.utils.vividict import Vividict
 logger = log.get_logger(__name__)
 
 
-class StateMachineModel(ModelMT, Hashable):
+class StateMachineModel(MetaModel, Hashable):
     """This model class manages a :class:`rafcon.core.state_machine.StateMachine`
 
     The model class is part of the MVC architecture. It holds the data to be shown (in this case a state machine).
@@ -63,7 +64,7 @@ class StateMachineModel(ModelMT, Hashable):
     def __init__(self, state_machine, meta=None, load_meta_data=True):
         """Constructor
         """
-        ModelMT.__init__(self)  # pass columns as separate parameters
+        MetaModel.__init__(self)  # pass columns as separate parameters
 
         assert isinstance(state_machine, StateMachine)
 
@@ -164,6 +165,10 @@ class StateMachineModel(ModelMT, Hashable):
     def update_hash(self, obj_hash):
         self.update_hash_from_dict(obj_hash, self.root_state)
         self.update_hash_from_dict(obj_hash, self.meta)
+
+    def update_meta_data_hash(self, obj_hash):
+        super(StateMachineModel, self).update_meta_data_hash(obj_hash)
+        self.root_state.update_meta_data_hash(obj_hash)
 
     @ModelMT.observe("state", before=True)
     @ModelMT.observe("outcomes", before=True)